ai_engine/requirements.txt
spoon-ai-sdk      # core framework
spoon-toolkits      # extended 
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=requirements.txt pyproject.toml
#
aiohappyeyeballs>=2.4.4
    # via aiohttp
aiohttp>=3.13.2
aiosignal>=1.3.2
    # via aiohttp
annotated-types>=0.7.0
    # via pydantic
anyio>=4.8.0
    # via
    #   httpx
    #   openai
    #   starlette
attrs>=25.1.0
    # via aiohttp
certifi>=2025.1.31
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer>=3.4.1
    # via requests
distro>=1.9.0
    # via openai
fastapi>=0.115.7
    # via spoon-ai-sdk (pyproject.toml)
frozenlist>=1.5.0
    # via
    #   aiohttp
    #   aiosignal
greenlet>=3.1.1
    # via sqlalchemy
h11>=0.14.0
    # via httpcore
httpcore>=1.0.7
    # via httpx
httpx>=0.28.1
    # via
    #   langsmith
    #   openai
idna>=3.10
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
jiter>=0.5.0
    # via openai
jsonpointer>=3.0.0
    # via jsonpatch
multidict>=6.1.0
    # via
    #   aiohttp
    #   yarl
neo-mamba>=3.0.1
openai>=1.70.0
orjson>=3.10.15
    # via langsmith

propcache>=0.2.1
    # via yarl
pydantic>=2.10.4
    # via
    #   langsmith
    #   openai
    #   spoon-ai-sdk (pyproject.toml)
pydantic-core>=2.27.2
    # via pydantic
python-dotenv>=1.0.1
    # via spoon-ai-sdk (pyproject.toml)
regex>=2024.11.6
    # via tiktoken
requests>=2.32.3
    # via
    #   langsmith
    #   requests-toolbelt
    #   spoon-ai-sdk (pyproject.toml)
    #   tiktoken
requests-toolbelt>=1.0.0
    # via langsmith
sniffio>=1.3.1
    # via
    #   anyio
    #   openai
sqlalchemy>=2.0.38
starlette>=0.45.3
    # via fastapi
tenacity>=8.5.0
    # via
tiktoken>=0.8.0
tqdm>=4.67.1
    # via openai
typing-extensions>=4.12.2
    # via
    #   anyio
    #   fastapi
    #   openai
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
urllib3>=2.3.0
    # via requests
yarl>=1.18.3
    # via aiohttp
prompt_toolkit>=3.0.51

websockets==15.0.1

termcolor>=3.0.1
pinecone>=6.0.2
google>=3.0.0
google-genai>=1.11.0
protobuf>=3.19.5
google-api-core>=2.24.2
grpcio>=1.71.0
web3==7.11.0
# solana==0.35.1  # Conflicts with fastmcp websockets requirement
# solathon>=1.0.0  # Alternative Solana SDK with more flexible dependencies
nest_asyncio>=1.6.0
python-telegram-bot>=22.0
anthropic>=0.42.0
boto3==1.35.99
botocore==1.35.99

# Additional dependencies used in the core codebase
cryptography>=41.0.0
eth-utils>=2.0.0
rlp>=3.0.0
googleapis-common-protos>=1.60.0

fastmcp>=2.12.0

google-generativeai==0.8.5

uvicorn[standard]>=0.32.0

x402>=0.2.1
    # via spoon-ai-sdk (pyproject.toml)

# Extra dependencies merged from the separate Crypto PowerData requirements.txt
pandas>=1.0.0

# Crypto PowerData dependencies
ccxt>=4.0.0
numpy>=1.20.0
TA-Lib>=0.4.25
asyncio-throttle>=1.0.0
pydantic-settings>=2.0.0
gql>=3.5.0


# Development dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
bump2version>=1.0.0
pycares>=4.9,<5.0
